make programmer code more portable

+++ software/firmware/pru0-programmer/include/device.h
@@ -18,7 +18,7 @@ typedef enum
     DRV_ERR_PROTECTED = 1,
 } drv_err_t;
 
-typedef int (*fn_open_t)(unsigned int, unsigned int, unsigned int);
+typedef int (*fn_open_t)(uint8_t, uint8_t, uint32_t);
 typedef int (*fn_erase_t)(void);
 typedef int (*fn_read_t)(uint32_t *dst, uint32_t address);
 typedef int (*fn_write_t)(uint32_t address, uint32_t data);
@@ -34,7 +34,7 @@ typedef struct
     fn_verify_t  verify;
     fn_close_t   close;
     /* processor word width */
-    unsigned int word_width_bytes;
+    uint8_t      word_width_bytes;
 } device_driver_t;
 
 extern device_driver_t nrf52_driver;
